clip/requirements.txt
grpcio
grpcio-tools
torch
Pillow
git+https://github.com/openai/CLIP.git
